Fix integer overflow soundness in preservation traits and arithmetic
The preservation trait system incorrectly claimed that predicates like
Positive are preserved under addition/multiplication for all types,
including integers where overflow can silently violate the invariant.
This commit restricts preservation claims to floating-point types only,
adds overflow checks to abs() and square() for integers, removes dead
code, fixes the is_refined concept for capturing lambdas, and adds
comprehensive test coverage for previously-untested composition features.

Critical fixes:
- preserves<Pred, Op> now takes a type parameter T; specializations
  restricted to std::floating_point<T> so integer arithmetic returns
  std::optional (runtime-checked) while float returns Refined (zero-cost)
- abs() throws refinement_error on INT_MIN instead of invoking UB
- square() throws refinement_error on integer overflow
- Same-predicate operators exclude interval types via has_interval_bounds

Other fixes:
- is_refined concept checks predicate invocability (works with InRange)
- refinement_error constructor constrained with std::formattable
- IsUnit predicate made type-generic (was capturing double bounds)
- PowerOfTwo constrained to integral types
- Removed vacuous Index type (size_t is always non-negative)
- Removed dead code: constexpr_predicate_for, detail helpers, compose
  namespace operators, DEFINE_REFINED_TYPE macro
- Removed misleading requires clause on division operator
- Added sizeof static_asserts for zero-overhead guarantee
- Added safe_divide limitation comment about inf/inf producing NaN

+++ examples/zero_overhead/02_arithmetic.cpp
@@ -1,27 +1,33 @@
-// 02_arithmetic.cpp — Proves Positive + Positive == int + int
+// 02_arithmetic.cpp — Proves Positive<double> + Positive<double> == double + double
 //
-// The preserves<Positive, plus> trait lets operator+ use assume_valid,
-// so the addition should compile to a single `add` with no branches.
+// The preserves<Positive, plus, double> trait lets operator+ use assume_valid
+// for floating-point types, so the addition compiles to a single `addsd`
+// with no branches.
+//
+// For integers, overflow can violate Positive, so operator+ returns
+// std::optional and includes a runtime check (branch).
 
 #include <rcpp/refine.hpp>
 
 using namespace refine;
 
+// --- Floating-point: zero overhead ---
+
 __attribute__((noinline))
-int refined_add(Refined<int, Positive> a, Refined<int, Positive> b) {
+double refined_add(Refined<double, Positive> a, Refined<double, Positive> b) {
     return (a + b).get();
 }
 
 __attribute__((noinline))
-int plain_add(int a, int b) {
+double plain_add(double a, double b) {
     return a + b;
 }
 
 int main() {
-    auto a = Refined<int, Positive>(10, assume_valid);
-    auto b = Refined<int, Positive>(20, assume_valid);
-    volatile int sink;
+    auto a = Refined<double, Positive>(10.0, assume_valid);
+    auto b = Refined<double, Positive>(20.0, assume_valid);
+    volatile double sink;
     sink = refined_add(a, b);
-    sink = plain_add(10, 20);
+    sink = plain_add(10.0, 20.0);
     return 0;
 }
